+++ engine/lite/request-response/runtime/src/test/scala/pl/touk/nussknacker/engine/requestresponse/deployment/FileProcessRepositoryTest.scala
@@ -13,7 +13,7 @@ class FileProcessRepositoryTest extends FunSuite with Matchers {
 
   private val processId = 54544L
 
-  private val processJson = """{\"additionalBranches\":[],\"nodes\":[{\"ref\":{\"parameters\":[],\"typ\":\"request1-post-source\"},\"id\":\"start\",\"type\":\"Source\"},{\"ref\":{\"parameters\":[],\"typ\":\"response-sink\"},\"id\":\"endNodeIID\",\"type\":\"Sink\"}],\"metaData\":{\"typeSpecificData\":{\"path\":\"alamakota\",\"type\":\"StandaloneMetaData\"},\"id\":\"process1\"}}"""
+  private val processJson = """{\"additionalBranches\":[],\"nodes\":[{\"ref\":{\"parameters\":[],\"typ\":\"request1-post-source\"},\"id\":\"start\",\"type\":\"Source\"},{\"ref\":{\"parameters\":[],\"typ\":\"response-sink\"},\"id\":\"endNodeIID\",\"type\":\"Sink\"}],\"metaData\":{\"typeSpecificData\":{\"path\":\"alamakota\",\"type\":\"RequestResponseMetaData\"},\"id\":\"process1\"}}"""
   private val deploymentJson =
     s"""
       |{