+++ ui/server/src/main/scala/db/migration/postgres/V1_032__StandaloneToRequestResponse.scala
@@ -1,0 +1,8 @@
+package db.migration.postgres
+
+import db.migration.{V1_032__StandaloneToRequestResponseDefinition => V1_031__FragmentSpecificDataDefinition}
+import slick.jdbc.{JdbcProfile, PostgresProfile}
+
+class V1_032__StandaloneToRequestResponse extends V1_031__FragmentSpecificDataDefinition {
+  override protected lazy val profile: JdbcProfile = PostgresProfile
+}